Restructure internals dir
