+++ src/libs/Paidy/Internals/Text/Json/JsonConverterAttribute.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Text.Json.Serialization;
 
-namespace Paidy.Internals;
+namespace Paidy.Internals.Text.Json;
